Página "Mapa da Força" com dados provenientes da API do RsbWebService (closes #7)
 -> Trabalho realizado:
    - Criação de classes na camada de acesso a dados que tiram partido da biblioteca 'Async Http Client' (definida na classe 'RsbWebService') para realizar pedidos assíncronos aos recursos da API, nomeadamente os recursos 'pessoal' e 'presenças' com as classes 'PessoalWebService' e 'PresencasWebService' respectivamente;
    - Adaptação e melhorias das classes já existentes com base no ponto anterior.

+++ RSB-WebClient/src/main/java/pt/isel/ps/li61n/model/dal/IPessoalRepository.java
@@ -9,5 +9,5 @@ import pt.isel.ps.li61n.model.entities.Pessoal;
  * @author Carlos Marques - carlosmmarques@gmail.com
  *         Tiago Venturinha - tventurinha@gmail.com
  */
-public interface IPessoalRepository extends IRepository< Integer, Pessoal> {
+public interface IPessoalRepository extends IRepository< Long, Pessoal> {
 }
